requirements.txt
pytest
pyyaml
pandas
speedtest-cli
python-dotenv
requests
